+++ tests/Pest.php
@@ -1,0 +1,5 @@
+<?php
+
+use Emuniq\FilamentBrowserNotifications\Tests\TestCase;
+
+uses(TestCase::class)->in('Feature');